DashSql/viz/requirements.txt
streamlit
pandas
mysql-connector-python
plotly
networkx
matplotlib
tqdm
